+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Added TokenTickerIcon optional param to TokenSelect](https://github.com/multiversx/mx-sdk-dapp-form/pull/176)
+
 ## [[0.6.1](https://github.com/multiversx/mx-sdk-dapp-form/pull/174)] - 2023-05-19
 
 - [Added a new Receiver component with autocomplete support](https://github.com/multiversx/mx-sdk-dapp-form/pull/163)